+++ src/user/user.module.ts
@@ -2,15 +2,15 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { User } from '@/user/infra/entities';
-import { ListUserUseCase } from '@/user/use-cases';
-import { IUserRepository } from '@/user/infra/interfaces';
+import { ListUserByIdUseCase } from '@/user/use-cases';
+import { IUserRepository } from '@/user/interfaces';
 import { UserRepository } from '@/user/infra/repositories';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
   controllers: [],
   providers: [
-    ListUserUseCase,
+    ListUserByIdUseCase,
     {
       provide: IUserRepository,
       useClass: UserRepository,
